requirements.txt
beautifulsoup4==4.12.2
certifi==2023.11.7
Flask==3.0.0
Flask-Cors==4.0.0
google-api-python-client==2.113.0
numpy==1.26.2
pandas==2.1.4
pickle-mixin==1.0.2
pyOpenSSL==23.3.0
python-dateutil==2.8.2
python-whois==0.8.0
requests==2.31.0
tldextract==5.1.1
urllib3==2.1.0
xgboost==2.0.3
